app/requirements.txt
cassandra-driver
pathvalidate